requirements.txt
# Core Python Dependencies
python>=3.11

# Azure Services
azure-functions>=1.18.0
azure-storage-blob>=12.19.0
azure-servicebus>=7.11.0
azure-identity>=1.15.0
azure-keyvault-secrets>=4.7.0
azure-monitor-opentelemetry>=1.2.0

# Machine Learning & AI
torch>=2.1.0
torchaudio>=2.1.0
openai-whisper>=20231117
whisperx>=3.1.1
transformers>=4.36.0
accelerate>=0.25.0

# Natural Language Processing
spacy>=3.7.0
spacy[pt]>=3.7.0

# Audio Processing
librosa>=0.10.1
soundfile>=0.12.1
pydub>=0.25.1
webrtcvad>=2.0.10

# OpenTelemetry Instrumentation
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-instrumentation>=0.42b0
opentelemetry-instrumentation-azure-functions>=0.42b0
opentelemetry-instrumentation-requests>=0.42b0
opentelemetry-instrumentation-logging>=0.42b0
opentelemetry-exporter-otlp>=1.21.0
opentelemetry-exporter-otlp-proto-grpc>=1.21.0

# Logging & Monitoring
structlog>=23.2.0
python-json-logger>=2.0.7

# Web Framework & API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0

# Async & Concurrency
asyncio-mqtt>=0.16.1
aiofiles>=23.2.1
aiohttp>=3.9.0

# Data Processing
numpy>=1.24.0
pandas>=2.1.0
scipy>=1.11.0

# Caching & Performance
redis>=5.0.1
hiredis>=2.2.3

# Security & Cryptography
cryptography>=41.0.0
python-multipart>=0.0.6

# Utilities
python-dotenv>=1.0.0
typer>=0.9.0
click>=8.1.7
pyyaml>=6.0.1
pathlib2>=2.3.7
psutil>=5.9.6

# Testing Dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
httpx>=0.25.0

# Code Quality
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.0
bandit>=1.7.5
pre-commit>=3.6.0

# Development
ipython>=8.17.0
jupyter>=1.0.0